added code for leaderboard page still  needs editing and added 4 games for the games. also added to game dashboard as well
