--- client/src/pages/GameDashboardPage.jsx
@@ -1,1 +1,0 @@
-
