+++ client/src/pages/css/LeaderboardPage.css
@@ -1,0 +1,2 @@
+
+
